backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.20.0
atpublic==5.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.50.0
debugpy==1.8.11
decorator==5.2.1
Deprecated==1.2.18
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.1
dvc-data==3.16.9
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gto==1.7.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kombu==5.5.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
nest_asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
omegaconf==2.3.0
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peewee==3.17.9
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==5.8.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
pygit2==1.15.1
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.13.1
scmrepo==3.3.10
semver==3.0.4
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
sqltrie==0.11.2
stack_data==0.6.3
starlette==0.46.2
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.56
zc.lockfile==3.0.post1
zipp==3.21.0
prometheus-fastapi-instrumentator
